button component moved to primitives and accessible from components directory
