+++ app/components/layout/content-actionbar.tsx
@@ -1,7 +1,7 @@
 import { Link } from '@remix-run/react'
+import { Button } from '#app/components/index.ts'
 import { DeleteNote } from '#app/routes/users+/$username_+/notes.$noteId.tsx'
 import { floatingToolbarClassName } from '../floating-toolbar.tsx'
-import { Button } from '../ui/button.tsx'
 import { Icon, type IconName } from '../ui/icon.tsx'
 
 interface ContentActionbarProps {
