+++ app/components/ui/primitives/index.ts
@@ -1,0 +1,1 @@
+export * from './button.tsx'
