+++ app/components/ui/status-button.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { useSpinDelay } from 'spin-delay'
+import { Button, type ButtonProps } from '#app/components/index.ts'
 import { cn } from '#app/utils/misc.tsx'
-import { Button, type ButtonProps } from './button.tsx'
 import { Icon } from './icon.tsx'
 import {
 	Tooltip,
